fix .travis.sh for downstream test
 - add __init__.py to hrpsys_ros_bridge ( this should be fixed in current repo)
 - patch https://github.com/start-jsk/rtmros_common/pull/447 (stable rtc)
 - patch https://github.com/k-okada/hrpsys-base/commit/9ce00db.diff (stable rtc)

+++ .travis.sh
@@ -40,8 +40,52 @@ case $TEST_PACKAGE in
 
         sudo apt-get install -qq -y ros-hydro-$pkg
 
+        sudo apt-get install -qq -y ros-hydro-rqt-robot-dashboard
+
         # this is hotfix
-        sudo wget https://raw.githubusercontent.com/start-jsk/rtmros_common/master/hrpsys_tools/test/test-pa10.test -O /opt/ros/hydro/share/hrpsys_tools/test/test-pa10.test
+        if [ -e /opt/ros/hydro/share/hrpsys_tools ] ; then
+            sudo wget https://raw.githubusercontent.com/start-jsk/rtmros_common/master/hrpsys_tools/test/test-pa10.test -O /opt/ros/hydro/share/hrpsys_tools/test/test-pa10.test
+        fi
+
+        if [ -e /opt/ros/hydro/lib/python2.7/dist-packages/hrpsys_ros_bridge/ ] ; then
+            sudo touch /opt/ros/hydro/lib/python2.7/dist-packages/hrpsys_ros_bridge/__init__.py;
+
+            #
+            sudo sed -i s@imu_floor@odom@g /opt/ros/hydro/share/hrpsys_ros_bridge/test/test-samplerobot.py
+
+            #
+            sudo patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py <<EOF
+--- /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py 2014-04-17 17:28:42.000000000 +0900
++++ /opt/ros/hydro/share/hrpsys_ros_bridge/scripts/sensor_ros_bridge_connect.py 2014-04-28 00:30:27.250839313 +0900
+@@ -21,7 +21,7 @@
+             if rh.port(sen.name) != None: # check existence of sensor ;; currently original HRP4C.xml has different naming rule of gsensor and gyrometer
+                 print program_name, "connect ", sen.name, rh.port(sen.name).get_port_profile().name, bridge.port(sen.name).get_port_profile().name
+                 connectPorts(rh.port(sen.name), bridge.port(sen.name), "new")
+-                if sen.type == 'Force':
++                if sen.type == 'Force' and afs != None:
+                     print program_name, "connect ", sen.name, afs.port("off_" + sen.name).get_port_profile().name, bridge.port("off_" + sen.name).get_port_profile().name
+                     connectPorts(afs.port("off_" + sen.name), bridge.port("off_" + sen.name), "new") # for abs forces
+         else:
+
+EOF
+
+            sudo patch -p0 /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch <<EOF
+--- /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch 2014-04-17 17:28:42.000000000 +0900
++++ /opt/ros/hydro/share/hrpsys_ros_bridge/launch/hrpsys_ros_bridge.launch 2014-04-28 00:30:27.250839313 +0900
+@@ -10,9 +10,9 @@
+   <!-- Set these values false when using HIRO -->
+   <arg name="USE_COMMON" default="true" />
+   <arg name="USE_ROBOTHARDWARE" default="false" />
+-  <arg name="USE_WALKING" default="true" />
++  <arg name="USE_WALKING" default="false" />
+   <arg name="USE_COLLISIONCHECK" default="true" />
+-  <arg name="USE_IMPEDANCECONTROLLER" default="true" />
++  <arg name="USE_IMPEDANCECONTROLLER" default="false" />
+   <arg name="USE_GRASPCONTROLLER" default="false" />
+   <arg name="USE_TORQUECONTROLLER" default="false" />
+   <arg name="USE_SOFTERRORLIMIT" default="true" />
+EOF
+        fi
         #
         source /opt/ros/hydro/setup.bash
 
@@ -70,6 +114,7 @@ case $TEST_PACKAGE in
             sudo dpkg -r --force-depends ros-hydro-hrpsys
             catkin_make -j8 -l8
             catkin_make install -j8 -l8
+            source install/setup.bash
         else
             echo "
             #
@@ -82,13 +127,19 @@ case $TEST_PACKAGE in
             wstool set rtmros_hironx http://github.com/start-jsk/rtmros_hironx --git -y
             wstool set rtmros_nextage http://github.com/tork-a/rtmros_nextage --git -y
             wstool update
+
+            ## HOTFIX: https://github.com/start-jsk/rtmros_common/pull/447
+            wget https://github.com/start-jsk/rtmros_common/pull/447.diff
+            (cd rtmros_common; patch -p1 < ../447.diff)
             cd ..
             # do not copile hrpsys because we wan to use them
             sed -i "1imacro(dummy_install)\nmessage(\"install(\${ARGN})\")\nendmacro()" src/hrpsys/CMakeLists.txt
-            sed -i "s@install\(@dummy_install\(@f" src/hrpsys/CMakeLists.txt
-            sed -i "s@install\(@dummy_install\(@f" src/hrpsys/catkin.cmake
+            sed -i "s@install(@dummy_install(@g" src/hrpsys/CMakeLists.txt
+            sed -i "s@install(@dummy_install(@g" src/hrpsys/catkin.cmake
             catkin_make -j8 -l8 --only-pkg-with-deps `echo $pkg | sed s/-/_/g`
             catkin_make install -j8 -l8
+            rm -fr install/share/hrpsys ./install/lib/pkgconfig/hrpsys.pc
+            source install/setup.bash
 
             # checkokut old hrpsys
             mkdir -p ~/hrpsys_ws/src
@@ -100,12 +151,18 @@ case $TEST_PACKAGE in
 
             catkin_make -j8 -l8
             catkin_make install -j8 -l8
+            # HOTFIX: https://github.com/k-okada/hrpsys-base/commit/9ce00db.diff
+            sed -i "s@\['vs@#\['vs@g" install/lib/python2.7/dist-packages/hrpsys/hrpsys_config.py
+            sed -i "s@\['afs@#\['afs@g" install/lib/python2.7/dist-packages/hrpsys/hrpsys_config.py
+            sed -i "s@\['abc@#\['abc@g" install/lib/python2.7/dist-packages/hrpsys/hrpsys_config.py
+            sed -i "s@\['st@#\['st@g" install/lib/python2.7/dist-packages/hrpsys/hrpsys_config.py
+
             source install/setup.bash
+
             cd ~/catkin_ws
         fi
 
         rospack profile
-        source install/setup.bash
 
         export EXIT_STATUS=0;
         pkg_path=`rospack find \`echo $pkg | sed s/-/_/g\``
